extracted getDatabaseInfo() and updateInfo()
