+++ lib/appFunctions.js
@@ -6,13 +6,8 @@ const prompts = require('./prompts');
 const queries = require('./queries');
 const app = require('../app');
 
-const roleChoices = [];
-const employeeChoices = [];
-const departmentChoices = [];
-const employeeIds = [];
-
 function prompt() {
-  updateInfo(getDatabaseInfo);
+  prompts.initPrompts();
 
   inquirer.prompt(prompts.prompts.main).then(answer => {
     switch (answer.action) {
@@ -53,7 +48,6 @@ function prompt() {
         inquirer.prompt(prompts.prompts.add).then(answer => {
           switch (answer.Add) {
             case 'Employee':
-              // addEmployee();
               add('employee');
               break;
             case 'Department':
@@ -91,13 +85,13 @@ function prompt() {
         inquirer.prompt(prompts.prompts.delete).then(answer => {
           switch (answer.Remove) {
             case 'Employee':
-              // need this
+              remove('employee');
               break;
             case 'Department':
-              // need this
+              remove('department');
               break;
             case 'Role':
-              // need this
+              remove('role');
               break;
             case 'Back':
               prompt();
@@ -120,45 +114,45 @@ function prompt() {
 function view(choice) {
   switch (choice) {
     case 'departments':
-      getQuery(queries.allDepartments);
+      getQuery('view', queries.allDepartments);
       break;
     case 'roles':
-      getQuery(queries.allRoles);
+      getQuery('view', queries.allRoles);
       break;
     case 'employees':
-      getQuery(queries.allEmployees);
+      getQuery('view', queries.allEmployees);
       break;
     case 'employeesDepartment':
       // prompt to choose department then pass answer to getQuery
-      getQuery(queries.employeesWhere, { id: 1 });
+      getQuery('view', queries.employeesWhere, { id: 1 });
       break;
     case 'employeesManager':
       // prompt to choose manager then pass answer to getQuery
-      getQuery(queries.employeesWhere, { manager_id: 1 });
+      getQuery('view', queries.employeesWhere, { manager_id: 1 });
       break;
     case 'managers':
-      getQuery(queries.allManagers);
+      getQuery('view', queries.allManagers);
       break;
     case 'budget':
       // prompt to choose department then pass answer to getQuery
-      getQuery(queries.budget, 'Sales');
+      getQuery('view', queries.budget, 'Sales');
       break;
     default:
-      throw new Error('Unexpected choice, resulting in an error.');
+      throw new Error('Unexpected choice resulting in an error.');
   }
 }
 
 function add(choice) {
   switch (choice) {
     case 'employee':
-      employeeChoices.push('NULL');
+      // employeeChoices.push('NULL');
 
       inquirer.prompt(prompts.prompts.addEmployee).then(answer => {
         let roleId = roleChoices.indexOf(answer.role) + 1;
-        let managerIdIndex = employeeChoices.indexOf(answer.manager);
+        // let managerIdIndex = employeeChoices.indexOf(answer.manager);
         let managerId = employeeIds[managerIdIndex];
 
-        getQuery(queries.addEmployee, {
+        getQuery('add', queries.addEmployee, {
           first_name: answer.firstName,
           last_name: answer.lastName,
           role_id: roleId,
@@ -168,14 +162,14 @@ function add(choice) {
       break;
     case 'department':
       inquirer.prompt(prompts.prompts.addDepartment).then(answer => {
-        getQuery(queries.addDepartment, { name: answer.departmentName });
+        getQuery('add', queries.addDepartment, { name: answer.departmentName });
       });
       break;
     case 'role':
       inquirer.prompt(prompts.prompts.addRole).then(answer => {
         let departmentId = departmentChoices.indexOf(answer.roleDepartment) + 1;
 
-        getQuery(queries.addRole, {
+        getQuery('add', queries.addRole, {
           title: answer.roleName,
           salary: answer.salary,
           department_id: departmentId
@@ -185,110 +179,117 @@ function add(choice) {
   }
 }
 
-function update(choice) {
+function remove(choice) {
   switch (choice) {
-    case 'Employee Role':
-      updateEmployeeRole();
+    case 'employee':
+      // prompt to choose employee
+      getQuery('remove', queries.removeEmployee);
       break;
-    case 'Employee Manager':
-      break;
-  }
-}
-
-function updateEmployeeRole() {
-  inquirer.prompt(prompts.prompts.updateEmployeeRole).then(answer => {
-    let employeeIdIndex = employeeChoices.indexOf(answer.updateEmployee);
-    let employeeId = employeeIds[employeeIdIndex];
-    let newRole = roleChoices.indexOf(answer.roleUpdate) + 1;
-
-    getQuery(queries.updateEmployeeRole, [
-      { role_id: newRole },
-      { id: employeeId }
-    ]);
-  });
-}
-
-function getQuery(type, setting) {
-  app.connection.query(type, setting, (err, res) => {
-    if (err) throw err;
-    console.log(res);
-    const table = cTable.getTable(res);
-    console.log('\n' + table);
-    prompt();
-  });
-}
-
-function getDatabaseInfo(table, column) {
-  let choiceArr;
-
-  switch (table) {
     case 'department':
-      choiceArr = departmentChoices;
+      // prompt to choose department
+      getQuery('remove', queries.removeDepartment);
       break;
     case 'role':
-      choiceArr = roleChoices;
-      break;
-    case 'employee':
-      choiceArr = employeeChoices;
+      // prompt to choose role
+      getQuery('remove', queries.removeRole);
       break;
     default:
-      throw new Error('Unknown table, unexpected case.');
-  }
-
-  if (table === 'employee' && column === 'full_name') {
-    employeeChoices.length = 0;
-
-    app.connection.query(queries.employeesFullName, (err, res) => {
-      if (err) throw err;
-      for (const employees of res) {
-        choiceArr.push(employees.full_name);
-      }
-    });
-  } else {
-    app.connection.query(`SELECT ${column} FROM ${table}`, (err, res) => {
-      if (err) throw err;
-
-      switch (column) {
-        case 'name':
-          departmentChoices.length = 0;
-
-          for (const department of res) {
-            choiceArr.push(department.name);
-          }
-          break;
-        case 'title':
-          roleChoices.length = 0;
-
-          for (const role of res) {
-            choiceArr.push(role.title);
-          }
-          break;
-        case 'id':
-          employeeIds.length = 0;
-
-          for (const ids of res) {
-            employeeIds.push(ids.id);
-          }
-          employeeIds.sort((a, b) => a - b);
-          break;
-        default:
-          throw new Error('Unknown column, unexpected case.');
-      }
-    });
+      throw new Error('Unexpected choice resulting in an error.');
   }
 }
 
-function updateInfo(callback) {
-  callback('role', 'title');
-  callback('department', 'name');
-  callback('employee', 'full_name');
-  callback('employee', 'id');
+// function update(prompt, query, column) {
+//   inquirer.prompt(prompt).then(answer => {
+//     let columnChoice;
+
+//     switch (column) {
+//       case 'role':
+//         columnChoice = answer.roleUpdate;
+//         break;
+//       case 'manager':
+//         columnChoice = answer.managerUpdate;
+//         break;
+//     }
+
+//     let employeeIdIndex = employeeChoices.indexOf(answer.updateEmployee);
+//     let employeeId = employeeIds[employeeIdIndex];
+
+//     let newInfo = roleChoices.indexOf(answer.roleUpdate) + 1;
+
+//     getQuery(query, [{ role_id: newRole }, { id: employeeId }]);
+//   });
+// }
+
+// function update(choice) {
+//   switch (choice) {
+//     case 'Employee Role':
+//       inquirer.prompt(prompts.prompts.updateEmployeeRole).then(answer => {
+//         let employeeIdIndex = employeeChoices.indexOf(answer.updateEmployee);
+//         let employeeId = employeeIds[employeeIdIndex];
+//         let newRole = roleChoices.indexOf(answer.roleUpdate) + 1;
+
+//         getQuery('update', queries.updateEmployee, [
+//           { role_id: newRole },
+//           { id: employeeId }
+//         ]);
+//       });
+//       break;
+//     case 'Employee Manager':
+//       inquirer.prompt(prompts.prompts.updateEmployeeRole).then(answer => {
+//         let employeeIdIndex = employeeChoices.indexOf(answer.updateEmployee);
+//         let employeeId = employeeIds[employeeIdIndex];
+//         let newManager = roleChoices.indexOf(answer.roleUpdate) + 1;
+
+//         getQuery('update', queries.updateEmployee, [
+//           { manager_id: newManager },
+//           { id: employeeId }
+//         ]);
+//       });
+//       break;
+//   }
+// }
+
+// function updateEmployeeRole() {
+//   inquirer.prompt(prompts.prompts.updateEmployeeRole).then(answer => {
+//     let employeeIdIndex = employeeChoices.indexOf(answer.updateEmployee);
+//     let employeeId = employeeIds[employeeIdIndex];
+//     let newRole = roleChoices.indexOf(answer.roleUpdate) + 1;
+
+//     getQuery('update', queries.updateEmployee, [
+//       { role_id: newRole },
+//       { id: employeeId }
+//     ]);
+//   });
+// }
+
+function getQuery(choice, type, setting) {
+  app.connection.query(type, setting, (err, res) => {
+    if (err) throw err;
+
+    switch (choice) {
+      case 'view':
+        const table = cTable.getTable(res);
+        console.log('\n' + table);
+        prompt();
+        break;
+      case 'add':
+        console.log('-- Added! --');
+        prompt();
+        break;
+      case 'remove':
+        console.log('-- Removed! --');
+        prompt();
+        break;
+      case 'update':
+        console.log('-- Updated! --');
+        prompt();
+        break;
+      default:
+        app.connection.end();
+    }
+  });
 }
 
 module.exports = {
-  prompt,
-  getDatabaseInfo,
-  employeeChoices,
-  roleChoices,
-  departmentChoices
+  prompt
 };
