+++ lib/prompts.js
@@ -1,6 +1,83 @@
 'use strict';
 
-const appFunctions = require('./appFunctions');
+const app = require('../app');
+const queries = require('./queries');
+
+const roleChoices = [];
+const employeeChoices = [];
+const departmentChoices = [];
+const employeeIds = [];
+
+function initPrompts() {
+  updateInfo(getDatabaseInfo);
+}
+
+function getDatabaseInfo(table, column) {
+  let choiceArr;
+
+  switch (table) {
+    case 'department':
+      choiceArr = departmentChoices;
+      break;
+    case 'role':
+      choiceArr = roleChoices;
+      break;
+    case 'employee':
+      choiceArr = employeeChoices;
+      break;
+    default:
+      throw new Error('Unknown table, unexpected case.');
+  }
+
+  if (table === 'employee' && column === 'full_name') {
+    employeeChoices.length = 0;
+
+    app.connection.query(queries.employeesFullName, (err, res) => {
+      if (err) throw err;
+      for (const employees of res) {
+        choiceArr.push(employees.full_name);
+      }
+    });
+  } else {
+    app.connection.query(`SELECT ${column} FROM ${table}`, (err, res) => {
+      if (err) throw err;
+
+      switch (column) {
+        case 'name':
+          departmentChoices.length = 0;
+
+          for (const department of res) {
+            choiceArr.push(department.name);
+          }
+          break;
+        case 'title':
+          roleChoices.length = 0;
+
+          for (const role of res) {
+            choiceArr.push(role.title);
+          }
+          break;
+        case 'id':
+          employeeIds.length = 0;
+
+          for (const ids of res) {
+            employeeIds.push(ids.id);
+          }
+          employeeIds.sort((a, b) => a - b);
+          break;
+        default:
+          throw new Error('Unknown column, unexpected case.');
+      }
+    });
+  }
+}
+
+function updateInfo(callback) {
+  callback('role', 'title');
+  callback('department', 'name');
+  callback('employee', 'full_name');
+  callback('employee', 'id');
+}
 
 const prompts = {
   main: {
@@ -76,13 +153,13 @@ const prompts = {
       name: 'role',
       type: 'rawlist',
       message: "What is the employee's role?",
-      choices: appFunctions.roleChoices
+      choices: roleChoices
     },
     {
       name: 'manager',
       type: 'rawlist',
       message: "Who is the employee's manager?",
-      choices: appFunctions.employeeChoices
+      choices: employeeChoices
     }
   ],
   addDepartment: [
@@ -133,7 +210,7 @@ const prompts = {
       name: 'roleDepartment',
       type: 'list',
       message: 'What department is this role in?',
-      choices: appFunctions.departmentChoices
+      choices: departmentChoices
     }
   ],
   updateEmployeeRole: [
@@ -141,18 +218,18 @@ const prompts = {
       name: 'updateEmployee',
       type: 'list',
       message: "Which employee's role would you like to update?",
-      choices: appFunctions.employeeChoices
+      choices: employeeChoices
     },
     {
       name: 'roleUpdate',
       type: 'rawlist',
       message: "What is the employee's new role?",
-      choices: appFunctions.roleChoices
+      choices: roleChoices
     }
-  ],
-
+  ]
 };
 
 module.exports = {
-  prompts: prompts
+  prompts: prompts,
+  initPrompts
 };
